Added CsvGen
CsvGen is a utility created to generate a CSV on a periodic basis.
The current implementation prints a line to the screen everytime a
value is updated. The desire is to have a efficient and threadsafe
manner of generating benchmarking data within an executable, though
the current implementation is neither of these yet.wq